prep/image_enhancement/requirements.txt

basicsr 
facexlib 
gfpgan 
torch 
torchvision 
pillow
numpy